updated cloth controller with commandline executable parameters
distractors still interfere with the setup and there is no easy way to handle this -- so, I am ignoring distractors for this version. Similarly, changing the room to 'tdw' or 'house' leads to weird renderings where the flex_solid objects disappear! So, for this version I am using only the 'box' room.